+++ apps/web/src/features/budgets/getMonthlyBudget/fetchEffectiveMonthlyBudget.ts
@@ -1,8 +1,8 @@
 import * as z from "zod"
 
-import { getSupabaseClient } from "../../lib/supabase"
-import type { MonthlyBudget, MonthlyBudgetRow } from "./types"
-import { parseIsoDateOnlyToLocalDate, toMonthEndIsoDate } from "./utils/month"
+import { getSupabaseClient } from "../../../lib/supabase"
+import type { MonthlyBudget, MonthlyBudgetRow } from "../types"
+import { parseIsoDateOnlyToLocalDate, toMonthEndIsoDate } from "../utils/month"
 
 const monthlyBudgetRowSchema = z.object({
   id: z.number(),
